Break out larger course navigation block
